First working version
Simplified to a single function that receives in-memory PGF image data,
does the decoding and returns an ImageData instance ready to use in
Javascript.

libpgf's GetBitmap sets the alpha channel to 0, but ImageData process it
as transparent. For now I added a hack after decoding to transform the
alpha channel.